+++ TaunoKarki/src/health.cpp
@@ -1,7 +1,7 @@
 #include "health.h"
 #include "gameobject.h"
 
-Health::Health(GameObject* owner, int max) : Component(owner), max(max), current(max)
+Health::Health(GameObject* owner, float max) : Component(owner), max(max), current(max)
 {
 }
 
@@ -13,7 +13,7 @@ void Health::update(float deltaTime)
 {
 }
 
-void Health::change(int amount)
+void Health::change(float amount)
 {
 	current += amount;
 	if (current > max) current = max;
